oh-my-bash: Remove ineffective "fpath" initialization
"fpath" array is the feature of Zsh, to which there are no equivalent
in Bash.  Also, the current initialization code for "fpath" is broken;
$fpath is not expanded to the array elements but to the 0-th element
in Bash, so that only the last two paths are retained in the "fpath"
array.  Actually, oh-my-bash currently does not have any "fpath"
functions, so we can just safely remove the initialization code for
"fpath".

In case that some external scripts use the existing functions for
"fpath" initialization, we move the functions to
"lib/omb-deprecated.sh".